backend/requirements.txt
google-adk>=1.0.0
litellm>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
pandas>=2.2.0
python-multipart>=0.0.18
